Fixed typo, reorganized tests to make it easier to add more
