+++ extract_msg_tests/__init__.py
@@ -1,0 +1,6 @@
+from .ole_writer_tests import OleWriterTests
+
+
+__all__ = [
+    'OleWriterTests',
+]
